Civ 1849 disposal hearing populate fields (1/2) (#986)
* prepopulate disposal hearing dates

* prepopulate disposal hearing fields

+++ src/main/java/uk/gov/hmcts/reform/civil/model/sdo/JudgementSum.java
@@ -11,10 +11,10 @@ import lombok.Data;
 public class JudgementSum {
 
     @JsonFormat(shape = JsonFormat.Shape.STRING)
-    private final Double judgementSum;
+    private final Integer judgementSum;
 
     @JsonCreator
-    public JudgementSum(@JsonProperty("judgementSum") Double judgementSum) {
+    public JudgementSum(@JsonProperty("judgementSum") Integer judgementSum) {
         this.judgementSum = judgementSum;
     }
 }
